Switch authorization to user attributes
This adds groups and three roles a user can have in the context of a group.
Plus one list of all possible groups.

On each operation we check if the user is logged in and has the required role.

This requires two user facing API changes:

* POST/PUT to the requests requires a "groupId" to be passed, to which the
  request will be assigned to.
* GET on the requests collection requires a "groupId" query parameter to
  fetch all requests that are assgined to that group.

This currently doesn't use the READER_METADATA/CONTENT split which will result
in limited data being returned when the user only has the metadata reader role.
This will be implemented next.