Fix type duplications between ddpersonaldata and zuse
Both specs are included, but they duplicate lots of types with the
same name, but with slightly different content.

This confused the php soap generator and made it only output the
zuse stuff most of the times.

Duplicate the classes manually now by running the generator with most of zuse removed
from the wsdl and make sure we prefer the DD namespace in all cases if possible.