Add a new "USER" role which gets checked at the start of all entry points
We currently only have roles for users in the context of a group,
but not one level up in the context of the application/api.

This means any authenticated user would hit the database, and after
that get denied once the group information is available.

Instead we add a USER role which doesn't have a subject, which gets
checked in the beginning, before we read anything from the DB.