Move high level vendo specifics into its own class
This still references the vendo class directly everywhere, but it makes
it clear that it is a service provider specific function. In case we get
a second provider we can convert this to an interface and abstract it away.

Besides moving the vendo code there there is a small potential behaviour change.
Up until now we only parsed the notify status in case of status: vendo-P6, but
we now just parse the status response in every case. In theory the response
should only contain the pdf/unclaimed info in the final status, so this shouldn't
make a difference.