Work around api-platform no longer respecting operation attributes
For some reason the legacy write/read no longer work. I looked into the code
doing this but it's not clear to me where there error is (the operation names don't
match).

Just use the alternative attributes which still work for now, and look into migrating
those entities to the new system instead.