Move the submit input validation to before we start submitting
There are certain inputs which we validate such as a request name
or a proper refrence number before we allow submitting a request.

This happend during the function doing the soap call which is after
the request is already in the submitted state, so too late to abort
the submit state change.

Instead move the checks into checkRequestReadyForSubmit() which is
called before all this.

Also make the reference number nullabel everywhere, meaning it isn't set,
to make the code handling it consistent.