doctrine: switch to datetime_immutable and date_immutable
Newer doctrine no longer auto converts between mutable and immutable types
so we have to decide on one and make sure only that type is passed to
doctrine.

Switch to datetime_immutable and date_immutable everywhere and in all setters
that take DateTimeInterface make sure to convert to DateTimeInterface before
forwarding to doctrine.