PreAddressingResponse: handle case where AddressingResults is missing
We just got a case where the response returned success, contained errors,
and didn't not contain a "AddressingResults" element, unlike when there
are no errors but it didn't find any results.

This is in theory valid according to the soap spec, so handle it properly
and treat it as if the pre-addressing didn't return any result since we
can't do anything about it.

Also add a unit test with the response we got, both the test the parsing,
and as an example for what real-life responses are possible.

This fixes a 500 in case the dual-delivery backend fails during pre-addressing.